Arabic Support and ISO-693-1 formatted localization lproj directories
The following change covers:  - Modifications to the lproj directories
following the ISO-693-1 format and accompanying project file
differences. - Modifications to FRFeedbackReporter to allow for a
programatic Display Name through the UI rather than using the
CFBundleExecutable.  Note: This was needed to allow for branding of the
application name.  - Definition of App.xib as a localizable file
allowing for testing of the localized framework through the test appl
Note: Localization from a Framework is only applied when the process for
the application including it also supports that localization otherwise
it will not apply the localization.