Fixed issue #25 by not trying to show UI upon non-main-thread exception
The previous behaviour of calling [NSThread exit] was no good because
the thread might be owned by GCD and killing it is inappropriate.
Instead, now do as Cocoa does by default anyway, and kill the process
if an exception occurs on a non-main thread.  Upon relaunch, the user
will still be able to submit a crash report with the backtrace of the
exception itself.